requirements.txt
nltk==3.3
pandas==0.23.0
Flask==1.0.2
beautifulsoup4==4.8.0
mysql_connector_repackaged==0.3.1
scikit_learn==0.21.3

